door_sprites.c
